changed float to decimal for Debit, Expense, Income schema. edited migrations also

+++ db/migrate/20230315135933_create_expenses.rb
@@ -2,7 +2,7 @@ class CreateExpenses < ActiveRecord::Migration[7.0]
   def change
     create_table :expenses do |t|
       t.string :name
-      t.float :balance
+      t.decimal :balance
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
