+++ db/migrate/20230315140540_create_incomes.rb
@@ -2,7 +2,7 @@ class CreateIncomes < ActiveRecord::Migration[7.0]
   def change
     create_table :incomes do |t|
       t.string :name
-      t.float :balance
+      t.decimal :balance
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
